backend/requirements.txt
django
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-simplejwt